clean up the file
